Restructure project layout
* Move all code into a "snekbox" package
* Use logging code as __init__.py
* Rename Snekbox class to NsJail
* Create "site" sub-package
    * Move templates into this sub-package
    * Move Flask code into a new snekapp module
