+++ .flake8
@@ -1,6 +1,6 @@
 [flake8]
 max-line-length=100
-application_import_names=snekbox,config,logs
+application_import_names=snekbox
 ignore=
     P102,B311,W503,E226,S311,
     # Missing Docstrings
